pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "django-quanttide"
version = "0.1.0-beta.1"
description = "量潮Django SDK"
authors = ["QuantTide Inc. <opensource@quanttide.com>"]
license = "Apache 2.0"
readme = "README.md"
packages = [{include = "django_quanttide"}]

[tool.poetry.dependencies]
python = "^3.8"
Django = ">=3.1"
djangorestframework = ">=3.0"
django-polymorphic = "^3.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"
coverage = "^7.2.7"
pre-commit = "^3.3.3"
tox = "^4.6.3"
